Update to 1.19.4

+++ src/main/java/exopandora/worldhandler/builder/argument/BlockPosArgument.java
@@ -72,7 +72,7 @@ public class BlockPosArgument implements IArgument
 			return null;
 		}
 		
-		return new BlockPos(this.x.getValue().doubleValue(), this.y.getValue().doubleValue(), this.z.getValue().doubleValue());
+		return new BlockPos(this.x.getValue().intValue(), this.y.getValue().intValue(), this.z.getValue().intValue());
 	}
 	
 	@Override
